set up ts-jest
