+++ jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+    "root": [
+        "<rootDir>/src"
+    ],
+    "testMatch": [
+        "**/__tests__/**/*.+(ts|js|tsx|jsx)"
+    ],
+    "transform": [
+        "^.+\\.(ts|tsx)$" : "ts-jest"
+    ]
+}
